+++ wow_message_parser/src/rust_printer/structs/print_common_impls/print_size.rs
@@ -56,7 +56,8 @@ pub(crate) fn print_size_of_ty_rust_view(s: &mut Writer, m: &RustMember, prefix:
             format!("{prefix}{name}.iter().fold(0, |acc, x| acc + x.size())",)
         }
 
-        RustType::VariableItemRandomProperty
+        RustType::CacheMask
+        | RustType::VariableItemRandomProperty
         | RustType::NamedGuid
         | RustType::EnchantMask
         | RustType::InspectTalentGearMask