+++ wow_message_parser/src/rust_printer/structs/print_common_impls/print_update_mask_struct/from_range.rs
@@ -90,7 +90,8 @@ pub(crate) fn create_from_range(s: &mut Writer, words: &[Vec<UpdateMaskMember>])
                         unimplemented!("{:#?}", member.ty())
                     }
 
-                    Type::Population
+                    Type::CacheMask
+                    | Type::Population
                     | Type::AddonArray
                     | Type::String
                     | Type::CString
@@ -152,7 +153,8 @@ pub(crate) fn create_from_range(s: &mut Writer, words: &[Vec<UpdateMaskMember>])
                             unimplemented!()
                         }
 
-                        Type::Population
+                        Type::CacheMask
+                        | Type::Population
                         | Type::AddonArray
                         | Type::String
                         | Type::CString