+++ wow_message_parser/src/rust_printer/structs/test_case_string/bytes.rs
@@ -310,7 +310,8 @@ fn print_bytes_definition(
             s.closing_curly(); // if !is empty
         }
 
-        Type::UpdateMask { .. }
+        Type::CacheMask
+        | Type::UpdateMask { .. }
         | Type::AuraMask
         | Type::AchievementDoneArray
         | Type::AchievementInProgressArray