+++ wow_message_parser/src/rust_printer/structs/test_case_string/members.rs
@@ -357,7 +357,8 @@ fn print_member_definition(
             test_case_string::wln(s, format!("{prefix}];"));
         }
 
-        Type::UpdateMask { .. }
+        Type::CacheMask
+        | Type::UpdateMask { .. }
         | Type::AuraMask
         | Type::AchievementDoneArray
         | Type::AchievementInProgressArray