+++ wow_message_parser/src/wireshark_printer/printer.rs
@@ -544,6 +544,9 @@ fn print_definition(
         Type::Population => {
             unreachable!("population only in login")
         }
+        Type::CacheMask => {
+            unreachable!("cache mask only in wrath")
+        }
     }
 }